+++ src/app/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { AuthModule as CoreAuthModule } from '@receipt-wrangler/receipt-wrangler-core';
 
 import { AuthRoutingModule } from './auth-routing.module';
-import { MobileAuthFormComponent } from './auth-form/mobile-auth-form.component';
+import { MobileAuthFormComponent } from './mobile-auth-form/mobile-auth-form.component';
 
 @NgModule({
   declarations: [MobileAuthFormComponent],
